pyproject.toml
[tool.poetry]
name = "arch_nemesis"
version = "0.1.0"
description = "AUR Package Updater"
authors = ["Dan Trickey <dan@trickey.io>"]
license = "MIT"
readme = "README.md"
repository = "https://github.com/"
homepage = "https://github.com/"
documentation = "https://readthedocs.io"
classifiers = [
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Typing :: Typed",
]

[tool.poetry.dependencies]
python = "^3.7"
pygithub = "^1.51"
pydantic = "^1.6.1"
"ruamel.yaml" = "^0.16.10"
click = "^7.1.2"
colorama = "^0.4.3"
cached_property = "^1.5.1"
requests = "^2.24.0"
jinja2 = "^2.11.2"
gitpython = "^3.1.7"

[tool.poetry.dev-dependencies]
flake8 = "*"
flake8-commas = "*"
flake8-comprehensions = "*"
flake8-debugger = "*"
flake8-docstrings = "*"
flake8-isort = "*"
flake8-mutable = "*"
flake8-todo = "*"
mypy = "*"
pytest = "*"
pytest-cov = "*"
rope = "*"
isort = {version = "*",extras = ["pyproject"]}


[tool.poetry.scripts]
arch-nemesis = 'arch_nemesis.app:main'

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
